🔨 [Refactor] GuardianProfile 모델 분리 및 적용
